Version 1.0
Added variable g:ttcn_minlines. If this variable is set (e.g. let g:ttcn_minlines=100) it will be used to set "minlines" in the syntax synchronization. If it is not set ":syn sync fromstart" will be used. Thanks to John Orr for pointing out a problem that led to this change.